backend/requirements.txt
Flask
Flask-Redis
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
flask-cors
Gunicorn